requirements.txt
requests==2.28.2
python-dotenv==1.0.0
pandas==2.0.0
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
sqlalchemy==2.0.9
